pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "nitrobox"
version = "0.0.5"
description = "Lightweight Linux namespace sandbox with persistent shell and instant reset"
requires-python = ">=3.12"
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Rust",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
]
readme = "README.md"
keywords = [
    "sandbox",
    "container",
    "namespace",
    "linux",
    "isolation",
    "reinforcement-learning",
    "overlayfs",
    "rootless",
    "checkpoint",
    "criu",
]
authors = [{ name = "Yuzhou Nie" }]
dependencies = ["protobuf>=4.0", "pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/opensage-agent/nitrobox"
Repository = "https://github.com/opensage-agent/nitrobox"
"Bug Tracker" = "https://github.com/opensage-agent/nitrobox/issues"
Changelog = "https://github.com/opensage-agent/nitrobox/blob/main/CHANGELOG.md"

[project.scripts]
nitrobox = "nitrobox.cli:main"
nbx = "nitrobox.cli:main"

[project.optional-dependencies]
dev = ["pytest>=9.0"]

[tool.maturin]
python-source = "src"
module-name = "nitrobox._core"
features = ["pyo3/extension-module"]
include = [
    { path = "src/nitrobox/_vendor/**", format = "sdist" },
]
